+++ src/main/java/com/consultrix/consultrixserver/repository/StudentRepository.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Long> {
-    List<Student> findCohortById(Long cohortId);
+public interface StudentRepository extends JpaRepository<Student, Integer> {
+    List<Student> findCohortById(Integer cohortId);
     List<Student> findByPipelineStage(String pipelineStage);
     List<Student> findByGraduationStatus(String graduationStatus);
 }
